Fixed all the comments, added a string generator to generate realistic values, added a getter for the value field to the Motif class.

+++ server/catgenome/src/main/java/com/epam/catgenome/entity/reference/motif/MotifSearchRequest.java
@@ -37,6 +37,6 @@ public class MotifSearchRequest {
     String chromosome;
     Integer startPosition;
     Integer endPosition;
-    Long pageSize;
+    Integer pageSize;
 
 }
