+++ server/catgenome/src/main/java/com/epam/catgenome/entity/reference/motif/MotifSearchResult.java
@@ -35,5 +35,5 @@ public class MotifSearchResult {
     List<Motif> result;
     String chr;
     Integer position;
-    Long pageSize;
+    Integer pageSize;
 }
